Split Linux memory policy processing into files
